% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCoSineSimilarity.R
\name{calculateCosineSimilarity}
\alias{calculateCosineSimilarity}
\title{Calculate Cosine Similarity Between Two Vectors}
\usage{
calculateCosineSimilarity(targetMean, comparatorMean)
}
\arguments{
\item{targetMean}{A numeric vector representing the first set of values.}

\item{comparatorMean}{A numeric vector representing the second set of values.}
}
\value{
A numeric value representing the cosine similarity between the two vectors.
}
\description{
This function calculates the cosine similarity between two numeric vectors.
Cosine similarity measures the cosine of the angle between two vectors,
which is a value between -1 and 1. A cosine similarity of 1 indicates that
the vectors are identical, 0 indicates orthogonality, and -1 indicates that
the vectors point in completely opposite directions.
}
