% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateStandardizedDifference.R
\name{calculateStandardizedDifference}
\alias{calculateStandardizedDifference}
\title{Calculate Standardized Difference Between Target and Comparator Proportions}
\usage{
calculateStandardizedDifference(targetProportion, comparatorProportion)
}
\arguments{
\item{targetProportion}{A data frame containing the target group's proportions with a column \code{targetMean}.}

\item{comparatorProportion}{A data frame containing the comparator group's proportions with a column \code{comparatorMean}.}
}
\value{
A data frame with the shared columns between the target and comparator data frames,
as well as the calculated standardized differences. The result includes columns for the
means (\code{targetMean}, \code{comparatorMean}), counts (if available), and the \code{stdDiff} column representing
the standardized difference.
}
\description{
This function calculates the standardized difference between two data frames containing
target and comparator proportions. It assumes that the input data frames have columns
\code{targetMean} and \code{comparatorMean} representing the proportions of the target and comparator groups, respectively.
The standardized difference is computed using the pooled standard deviation of the proportions.
}
\details{
The function performs the following steps:
\enumerate{
\item Ensures that the required columns \code{targetMean} and \code{comparatorMean} exist in the provided data frames.
\item Identifies shared columns between the two data frames and joins them.
\item Computes the standard deviation for the target and comparator proportions using the formula:
\deqn{SD = \sqrt{p(1 - p)}}
where \code{p} is the proportion.
\item Pooled standard deviation is calculated as:
\deqn{pooledSD = \sqrt{\frac{(SD_{Target}^2 + SD_{Comparator}^2)}{2}}}
\item The standardized difference is then computed as:
\deqn{stdDiff = \frac{comparatorMean - targetMean}{pooledSD}}
}
}
