% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessAnnualizedCrudeIncidenceRateData.R
\name{checkCrudeIncidenceRateDataStratifiedIsAnnualized}
\alias{checkCrudeIncidenceRateDataStratifiedIsAnnualized}
\title{Check If Data is Cohort Diagnostics Incidence Rate Data}
\usage{
checkCrudeIncidenceRateDataStratifiedIsAnnualized(data)
}
\arguments{
\item{data}{A data frame that is expected to contain the columns cohortCount, personYears, gender, ageGroup, calendarYear, incidenceRate, cohortId, and databaseId.}
}
\value{
This function does not return a value. It stops and throws an error if the input data does not meet the expected conditions.
}
\description{
This function checks if the input data is a data frame and contains all the expected columns with the correct data types for cohort diagnostics incidence rate data.
}
