% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessAnnualizedCrudeIncidenceRateData.R
\name{checkTemporalStabilityForannualizedCrudeIncidenceRateData}
\alias{checkTemporalStabilityForannualizedCrudeIncidenceRateData}
\title{Check Temporal Stability For Cohort Diagnostics Incidence Rate Data}
\usage{
checkTemporalStabilityForannualizedCrudeIncidenceRateData(
  annualizedCrudeIncidenceRateData,
  cohort,
  maxNumberOfSplines = NULL,
  splineTickInterval = 3,
  maxRatio = 1.25,
  alpha = 0.05,
  removeOutlierZeroCounts = TRUE
)
}
\arguments{
\item{annualizedCrudeIncidenceRateData}{A data frame that contains the columns cohortId, databaseId, gender, ageGroup, calendarYear, cohortCount, personYears, and incidenceRate.}

\item{cohort}{A data frame that contains the columns cohortId and cohortName.}

\item{maxNumberOfSplines}{The maximum number of splines to use in the model. If NULL, the default is 5.}

\item{splineTickInterval}{The interval at which to place the splines. The default is 3.}

\item{maxRatio}{The maximum ratio for the likelihood comparison. If NULL, the default is 1.25.}

\item{alpha}{The significance level for the likelihood ratio test. If NULL, the default is 0.05.}

\item{removeOutlierZeroCounts}{(Default TRUE) Do you want to remove continous zero counts in the beginning and end in the time series?}
}
\value{
A data frame with the cohortId, databaseId, gender, ageGroup, stable flag, and isUnstable flag.
}
\description{
This function checks the temporal stability of the cohort diagnostics incidence rate data. It processes the data, calculates the predicted counts, and compares the likelihoods.
}
