% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRelationship.R
\docType{package}
\name{CohortDiagnosticsLite-package}
\alias{CohortDiagnosticsLite}
\alias{CohortDiagnosticsLite-package}
\title{CohortDiagnosticsLite: Do CohortDiagnostics.}
\description{
Do CohortDiagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortDiagnosticsLite/}
  \item \url{https://github.com/OHDSI/CohortDiagnosticsLite}
  \item Report bugs at \url{https://github.com/OHDSI/CohortDiagnosticsLite/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
