% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareIndexEventBreakdown.R
\name{compareIndexEventBreakdown}
\alias{compareIndexEventBreakdown}
\title{Compare Index Event Breakdown between Target and Comparator Cohorts}
\usage{
compareIndexEventBreakdown(
  indexEventBreakdown,
  cohortCount,
  targetCohortId,
  targetDatabaseId = NULL,
  comparatorCohortId,
  comparatorDatabaseId = targetDatabaseId,
  compareEntries = FALSE
)
}
\arguments{
\item{indexEventBreakdown}{A data frame containing index event breakdown information. It should include fields such as
\code{cohortDefinitionId}, \code{persons}, and \code{records}, along with \code{databaseId} and other event-related details.}

\item{cohortCount}{A data frame containing cohort counts, including \code{cohortSubjects} and \code{cohortEntries} for each \code{cohortDefinitionId}.}

\item{targetCohortId}{An integer representing the cohort definition ID of the target cohort to be compared.}

\item{targetDatabaseId}{(Optional) A string representing the ID of the target database. If NULL, it will be derived from the \code{indexEventBreakdown}.}

\item{comparatorCohortId}{An integer representing the cohort definition ID of the comparator cohort to be compared.}

\item{comparatorDatabaseId}{(Optional) A string representing the ID of the comparator database. If NULL, it will default to \code{targetDatabaseId}.}

\item{compareEntries}{A boolean indicating whether to compare records (if \code{TRUE}) or persons (if \code{FALSE}). Defaults to \code{FALSE}.}
}
\value{
A data frame with the combined breakdown of index events for the target and comparator cohorts. It includes
fields such as cohort IDs, database IDs, event concepts, event counts, and computed means (proportions) for both
target and comparator cohorts. The function also returns standardized differences between the target and comparator cohorts.
}
\description{
This function compares the breakdown of index events between a target cohort and a comparator cohort.
It calculates the proportions of persons or records (depending on the \code{compareEntries} parameter) for each cohort
and combines the results into a single data frame. The function also computes standardized differences between
the target and comparator cohorts.
}
\details{
\itemize{
\item The \code{compareEntries} parameter determines whether the comparison is based on the number of \code{records} or \code{persons}.
If \code{compareEntries = TRUE}, the function compares records (from \code{cohortEntries} and \code{records}); otherwise, it compares persons
(from \code{cohortSubjects} and \code{persons}).
\item The \code{computeStandardizedDifference} function is applied to the proportions of both cohorts to calculate the standardized
differences.
}
}
