% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnnualizedCrudeIncidenceRate.R
\name{getAnnualizedCrudeIncidenceRate}
\alias{getAnnualizedCrudeIncidenceRate}
\title{Calculate Crude Annualized Incidence Rate}
\usage{
getAnnualizedCrudeIncidenceRate(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortDefinitionId,
  cdmDatabaseSchema,
  washoutPeriod = 365,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortTable,
  databaseId
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortDefinitionId}{The cohort definition ID for which the incidence rate is to be calculated.}

\item{cdmDatabaseSchema}{Schema name where your CDM tables reside.}

\item{washoutPeriod}{The washout period (in days) before a subject is eligible for their first event (default is 365 days).}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortTable}{The name of the cohort table.}

\item{databaseId}{A string name for the data source.}
}
\value{
A tibble containing the incidence rates for each calendar year. The tibble consists of:
\item{calendarYear}{The calendar year for which the incidence rate is calculated.}
\item{incidenceRate}{The calculated crude incidence rate.}
}
\description{
This function calculates the crude annualized incidence rate for a given cohort in a CDM (Common Data Model) database.
It returns the incidence rates based on calendar year for subjects in the specified cohort.
}
\details{
This function calculates the crude annualized incidence rate using the first occurrence of the event (if specified) and
a specified washout period. The incidence rate is computed based on the number of first occurrences within a cohort
for each calendar year.

SQL queries are executed to gather year range data, compute incidence rates, and return the summary.
Temporary tables are used during processing and are dropped after completion.
}
