% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GenerateCalendarYearCohortFromObservationPeriod.R
\name{getCalendarYearCohortFromObservationPeriod}
\alias{getCalendarYearCohortFromObservationPeriod}
\title{Creates a temporary table with calendar year cohorts from the observation period.}
\usage{
getCalendarYearCohortFromObservationPeriod(
  connection,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  tempCohortTableName = "#calendar_year_cohort",
  anchorDay = 1,
  anchorMonth = 1
)
}
\arguments{
\item{connection}{The connection object to the database.}

\item{cdmDatabaseSchema}{The schema name where the CDM (Common Data Model) tables are located.}

\item{tempEmulationSchema}{Schema to emulate temp tables in a non-temp schema. Defaults to the value of the \code{sqlRenderTempEmulationSchema} option.}

\item{tempCohortTableName}{The name of the temporary cohort table to be created. Default is \verb{#calendar_year_cohort}.}

\item{anchorDay}{The day of every month to anchor the cohort start date. Default is 1.}

\item{anchorMonth}{The month of the year to anchor the cohort start date. Default is 1.}
}
\value{
NULL The function does not return a value but creates a temporary cohort table in the database.
}
\description{
This function generates a calendar year cohort based on the observation period
for each person in the database. The cohort start date is anchored on a specified day
and month, and the resulting cohort is stored in a temporary table.
}
\details{
This function constructs SQL to create a calendar year cohort table. The cohort table will
be based on the observation period and anchored to the specified \code{anchorDay} and \code{anchorMonth}.
It uses the CDM database schema and translates the SQL for the target database platform.
}
