% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndexEventBreakdown.R
\name{getIndexEventBreakdown}
\alias{getIndexEventBreakdown}
\title{Get Index Event Breakdown}
\usage{
getIndexEventBreakdown(
  cohortIds,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{cohortIds}{A vector of cohort IDs to filter the cohort table for inclusion.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your CDM tables reside.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{The name of the cohort table.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}
}
\value{
A tibble containing the breakdown of counts for various clinical events. The tibble consists of:
\item{conceptId}{The concept ID of the event (e.g., visit, procedure, drug).}
\item{persons}{The number of distinct persons having the event.}
\item{records}{The total number of records of the event.}
\item{source}{The source of the event count, e.g., 'v1' for visit, 'v2' for visit source, 'p1' for procedure, etc.}
}
\description{
This function retrieves a breakdown of events (e.g., visits, procedures, drugs, observations, conditions, measurements)
associated with the index event (on first cohort start date of a subject) for a given set of cohorts.
}
\details{
For each event type (visit, procedure, drug exposure, observation, condition, and measurement), both standard
concept IDs and source concept IDs are retrieved, ensuring no duplication.

This function processes clinical event data for cohorts, such as visits, procedures, drugs, observations, conditions, and
measurements, based on the index date (cohort start date). It counts the number of persons and records for each event and
combines them into a single tibble, distinguishing between standard and source concepts.

SQL queries are translated and executed using the \code{DatabaseConnector} package. Temporary tables are dropped after processing.
}
