% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlotPopulationPyramid.R
\name{getPlotPopulationPyramid}
\alias{getPlotPopulationPyramid}
\title{Generate a Population Pyramid Plot}
\usage{
getPlotPopulationPyramid(data)
}
\arguments{
\item{data}{A tibble or data frame containing demographic information. It must have three
columns:
\describe{
\item{ageGroup}{Character vector representing the age group stratification (e.g., "0-9", "10-19").}
\item{gender}{Character vector representing the gender stratification ("Female" or "Male").}
\item{proportion}{Numeric vector representing the proportion of the population in each age group and gender.}
}}
}
\value{
A ggplot object displaying the population pyramid. The x-axis represents the age groups,
and the y-axis represents the proportion of the population (with male proportions shown as negative
values to the left and female proportions as positive values to the right). The fill color differentiates
between genders, where males are represented in blue and females in pink.
}
\description{
This function generates a population pyramid plot based on the provided demographic data.
It displays the proportion of males and females across different age groups, with male
proportions displayed on the left (as negative values) and female proportions on the right.
The plot is flipped to create the pyramid style.
}
\details{
The population pyramid visually represents the age and gender distribution of a population. It’s
a useful tool for understanding demographic structures. The function adjusts the male proportions
to be negative so that they appear on the left side of the plot, and it flips the coordinates to
achieve the pyramid-like shape.

The function requires the \code{ggplot2} package to create the plot. If \code{ggplot2} is not installed, the
function will stop and prompt the user to install it.
}
