% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredictedCount.R
\name{getPredictedCount}
\alias{getPredictedCount}
\title{Get Predicted Count in a Data Frame}
\usage{
getPredictedCount(
  data,
  timeSequenceField,
  countField,
  personTimeField,
  maxNumberOfSplines = NULL,
  splineTickInterval = 3,
  alpha = 0.05,
  maxRatio = 1.25
)
}
\arguments{
\item{data}{A data frame that contains the columns specified in the function parameters.}

\item{timeSequenceField}{The field in the data frame that represents the time sequence.}

\item{countField}{The field in the data frame that represents the count of events.}

\item{personTimeField}{The field in the data frame that represents the person time.}

\item{maxNumberOfSplines}{The maximum number of splines to use in the model. If NULL, the default is 3.}

\item{splineTickInterval}{The interval at which to place the splines. The default is 3.}

\item{alpha}{The significance level for the likelihood ratio test. If NULL, the default is 0.05.}

\item{maxRatio}{The maximum ratio for the likelihood comparison. If NULL, the default is 1.25.}
}
\value{
A data frame with the observed and expected counts.
}
\description{
This function predicts the count of events in a data frame using a Poisson regression model with natural splines. It groups the data by timeSequenceField and fills in missing counts with predicted values.
The function also adds a flag to indicate whether the data was imputed.
}
