% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredictedEventCounts.R
\name{getPredictedEventCounts}
\alias{getPredictedEventCounts}
\title{Get Predicted Count in a Data Frame}
\usage{
getPredictedEventCounts(
  data,
  timeSequenceField,
  countField,
  personTimeField,
  maxNumberOfSplines = NULL,
  splineTickInterval = 3,
  alpha = 0.05,
  maxRatio = 1.25
)
}
\arguments{
\item{data}{A data frame that contains the columns specified in the function parameters.}

\item{timeSequenceField}{The field in the data frame that represents the time sequence.}

\item{countField}{The field in the data frame that represents the count of events.}

\item{personTimeField}{The field in the data frame that represents the person time.}

\item{maxNumberOfSplines}{The maximum number of splines to use in the model. If NULL, the default is 3.}

\item{splineTickInterval}{The interval at which to place the splines. The default is 3.}

\item{alpha}{The significance level for statistical tests and confidence intervals (default is 0.05).}

\item{maxRatio}{The maximum ratio for the likelihood comparison. If NULL, the default is 1.25.}
}
\value{
A data frame with the observed and predicted counts from both \code{Cyclops} and \code{glm()}, along with diagnostics such as p-values, confidence intervals (for GLM), and stability indicators.
}
\description{
This function predicts the count of events in a data frame using Poisson regression models with natural splines. It fits the model using both the \code{Cyclops} package and base \code{glm()}.
The function models event counts based on the field specified in the \code{timeSequenceField} parameter and includes an offset for the field specified in \code{personTimeField} to account for different exposure times. It evaluates model stability through likelihood ratio tests and other diagnostics.
The function adds flags to indicate model stability and significance, with separate indicators for the results of both \code{Cyclops} and \code{glm()} fits.
}
\details{
\strong{Input Data Structure:}
The input \code{data} should be a data frame with columns that are dynamically specified via function parameters.
\itemize{
\item The \code{timeSequenceField} parameter represents a column containing the time sequence or index of the events (e.g., date or time intervals).
\item The \code{countField} parameter refers to the column representing the observed count of events during each time interval.
\item The \code{personTimeField} parameter points to the column representing the exposure time or population-at-risk during the corresponding time interval.
The \code{data} should not contain duplicate rows for the same value in the column specified by \code{timeSequenceField}, and both the columns for \code{countField} and \code{personTimeField} must not contain missing (\code{NA}) values.
}
}
