% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessAnnualizedCrudeIncidenceRateData.R
\name{processannualizedCrudeIncidenceRateData}
\alias{processannualizedCrudeIncidenceRateData}
\title{Process Crude Incidence Rate Data from Cohort Diagnostics}
\usage{
processannualizedCrudeIncidenceRateData(
  annualizedCrudeIncidenceRateData,
  removeOutlierZeroCounts = TRUE
)
}
\arguments{
\item{annualizedCrudeIncidenceRateData}{A data frame that contains the columns cohortId, databaseId, calendarYear, cohortCount, personYears, and incidenceRate.}

\item{removeOutlierZeroCounts}{(Default TRUE) Do you want to remove continuous zero counts in the beginning and end in the time series?}
}
\value{
A data frame with processed cohort diagnostics incidence rate data.
}
\description{
This function processes the cohort diagnostics incidence rate data. It filters the data based on certain conditions, calculates the incidence rate, and imputes missing years.
}
